@@ -1,77 +1,83 @@
 {
-  "name": "material-dashboard-angular",
-  "version": "2.2.0",
-  "license": "MIT",
-  "scripts": {
-    "ng": "ng",
-    "start": "ng serve",
-    "build": "ng build",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e",
-    "install:clean": "rm -rf node_modules/ && rm -rf package-lock.json && npm install && ng serve -o"
-  },
-  "engines": {
-    "node": "6.11.1",
-    "npm": "3.10.9"
-  },
-  "private": true,
-  "dependencies": {
-    "@agm/core": "1.0.0-beta.5",
-    "@angular/animations": "7.0.2",
-    "@angular/cdk": "7.0.2",
-    "@angular/common": "7.0.2",
-    "@angular/compiler": "7.0.2",
-    "@angular/core": "7.0.2",
-    "@angular/forms": "7.0.2",
-    "@angular/http": "7.0.2",
-    "@angular/material": "7.0.2",
-    "@angular/platform-browser": "7.0.2",
-    "@angular/platform-browser-dynamic": "7.0.2",
-    "@angular/platform-server": "7.0.2",
-    "@angular/router": "7.0.2",
-    "ajv": "6.4.0",
-    "arrive": "2.4.1",
-    "bootstrap": "4.1.0",
-    "bootstrap-material-design": "4.1.1",
-    "bootstrap-notify": "3.1.3",
-    "chartist": "0.11.0",
-    "classlist.js": "1.1.20150312",
-    "core-js": "2.4.1",
-    "express": "4.16.3",
-    "googleapis": "28.1.0",
-    "hammerjs": "2.0.8",
-    "jquery": "3.2.1",
-    "moment": "2.22.1",
-    "perfect-scrollbar": "1.1.0",
-    "popper.js": "1.14.3",
-    "rxjs": "6.3.3",
-    "rxjs-compat": "6.3.3",
-    "web-animations-js": "2.3.1",
-    "zone.js": "0.8.26"
-  },
-  "devDependencies": {
-    "@angular/cli": "6.0.3",
-    "@angular/compiler-cli": "7.0.2",
-    "@angular/language-service": "7.0.2",
-    "@types/bootstrap": "3.3.32",
-    "@types/chartist": "0.9.34",
-    "@types/googlemaps": "3.30.8",
-    "@types/jasmine": "2.5.38",
-    "@types/jquery": "1.10.31",
-    "@types/node": "6.0.73",
-    "codelyzer": "4.2.1",
-    "jasmine-core": "3.1.0",
-    "jasmine-spec-reporter": "4.2.1",
-    "karma": "2.0.0",
-    "karma-chrome-launcher": "2.2.0",
-    "karma-cli": "1.0.1",
-    "karma-coverage-istanbul-reporter": "1.4.2",
-    "karma-jasmine": "1.1.1",
-    "protractor": "5.3.1",
-    "ts-node": "5.0.1",
-    "tslint": "5.9.1",
-    "typescript": "3.1.6",
-    "@angular-devkit/build-angular": "~0.6.3"
-  }
+    "name": "base_proyectos_material",
+    "version": "2.2.0",
+    "license": "MIT",
+    "scripts": {
+        "ng": "ng",
+        "start": "ng serve",
+        "build": "ng build",
+        "test": "ng test",
+        "lint": "ng lint",
+        "e2e": "ng e2e",
+        "install:clean": "rm -rf node_modules/ && rm -rf package-lock.json && npm install && ng serve -o"
+    },
+    "engines": {
+        "node": "6.11.1",
+        "npm": "3.10.9"
+    },
+    "private": true,
+    "dependencies": {
+        "@agm/core": "1.0.0",
+        "@angular-redux/store": "^10.0.0",
+        "@angular/animations": "8.2.8",
+        "@angular/cdk": "8.2.1",
+        "@angular/common": "8.2.8",
+        "@angular/compiler": "8.2.8",
+        "@angular/core": "8.2.8",
+        "@angular/forms": "8.2.8",
+        "@angular/http": "7.2.15",
+        "@angular/material": "^8.2.1",
+        "@angular/material-moment-adapter": "^8.0.1",
+        "@angular/platform-browser": "8.2.8",
+        "@angular/platform-browser-dynamic": "8.2.8",
+        "@angular/platform-server": "8.2.8",
+        "@angular/router": "8.2.8",
+        "@ngx-loading-bar/core": "^4.2.0",
+        "@ngx-loading-bar/router": "^4.2.0",
+        "ajv": "6.10.2",
+        "arrive": "2.4.1",
+        "bootstrap": "4.3.1",
+        "bootstrap-material-design": "4.1.2",
+        "bootstrap-notify": "3.1.3",
+        "chartist": "0.11.4",
+        "classlist.js": "1.1.20150312",
+        "core-js": "3.2.1",
+        "express": "4.17.1",
+        "googleapis": "43.0.0",
+        "hammerjs": "2.0.8",
+        "jquery": "3.4.1",
+        "moment": "2.24.0",
+        "ng-redux": "^4.4.1",
+        "perfect-scrollbar": "1.4.0",
+        "popper.js": "1.15.0",
+        "redux": "^4.0.4",
+        "rxjs": "6.5.3",
+        "rxjs-compat": "6.5.3",
+        "web-animations-js": "2.3.2",
+        "zone.js": "0.10.2"
+    },
+    "devDependencies": {
+        "@angular-devkit/build-angular": "~0.803.6",
+        "@angular/cli": "8.3.6",
+        "@angular/compiler-cli": "8.2.8",
+        "@angular/language-service": "8.2.8",
+        "@types/bootstrap": "4.3.1",
+        "@types/chartist": "0.9.46",
+        "@types/googlemaps": "3.37.7",
+        "@types/jasmine": "3.4.1",
+        "@types/jquery": "3.3.31",
+        "@types/node": "12.7.8",
+        "codelyzer": "5.1.2",
+        "jasmine-core": "3.5.0",
+        "jasmine-spec-reporter": "4.2.1",
+        "karma": "4.3.0",
+        "karma-chrome-launcher": "3.1.0",
+        "karma-cli": "2.0.0",
+        "karma-coverage-istanbul-reporter": "2.1.0",
+        "karma-jasmine": "2.0.1",
+        "protractor": "5.4.2",
+        "ts-node": "8.4.1",
+        "tslint": "5.20.0",
+        "typescript": "~3.4.0"
+    }
 }
