@@ -3,27 +3,31 @@ import { NgModule } from '@angular/core';
 import { FormsModule,ReactiveFormsModule } from '@angular/forms';
 import { HttpModule } from '@angular/http';
 import { RouterModule } from '@angular/router';
-import { HttpClientModule } from '@angular/common/http';
+import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
 
-import { AppRoutingModule } from './app.routing';
+import { AppRoutingModule } from './app-routing.module';
 import { ComponentsModule } from './components/components.module';
 
 import { AppComponent } from './app.component';
 
-import { DashboardComponent } from './dashboard/dashboard.component';
-import { UserProfileComponent } from './user-profile/user-profile.component';
-import { TableListComponent } from './table-list/table-list.component';
-import { TypographyComponent } from './typography/typography.component';
-import { IconsComponent } from './icons/icons.component';
-import { MapsComponent } from './maps/maps.component';
-import { NotificationsComponent } from './notifications/notifications.component';
-import { UpgradeComponent } from './upgrade/upgrade.component';
-import {
-  AgmCoreModule
-} from '@agm/core';
+import { DashboardComponent } from './components/dashboard/dashboard.component';
+import {LoadingBarRouterModule} from '@ngx-loading-bar/router';
+
+
+
 import { AdminLayoutComponent } from './layouts/admin-layout/admin-layout.component';
-import { LoginComponent } from './login/login.component';
-import { HomeComponent } from './home/home.component';
+import { HomeComponent } from './components/home/home.component';
+
+// Interceptores
+import { JwtService } from './services/interceptors/jwt.service';
+import { ErrorsService } from './services/interceptors/errors.service';
+import { SidebarComponent} from '../app/layouts/components/sidebar/sidebar.component';
+import { NavbarComponent} from '../app/layouts/components/navbar/navbar.component';
+import { FooterComponent} from '../app/layouts/components/footer/footer.component';
+import { PagesLayoutComponent } from './layouts/pages-layout/pages-layout.component';
+import {LayoutModule} from '../app/layouts/components/layout.module';
+
+
 
 @NgModule({
   imports: [
@@ -31,23 +35,32 @@ import { HomeComponent } from './home/home.component';
     FormsModule,
     ReactiveFormsModule,
     HttpModule,
-    ComponentsModule,
     RouterModule,
     AppRoutingModule,
     HttpClientModule,
-    AgmCoreModule.forRoot({
-      apiKey: 'YOUR_GOOGLE_MAPS_API_KEY'
-    })
+    LoadingBarRouterModule,
+    
   ],
   declarations: [
     AppComponent,
     AdminLayoutComponent,
-    LoginComponent,
     HomeComponent,
+     SidebarComponent,
+    NavbarComponent,
+    FooterComponent,
+    PagesLayoutComponent
     
 
   ],
-  providers: [],
-  bootstrap: [AppComponent]
+  providers: [
+      {
+        provide: HTTP_INTERCEPTORS, useClass: JwtService, multi: true,
+      },
+      {
+        provide: HTTP_INTERCEPTORS, useClass: ErrorsService, multi: true,
+      }
+  ],
+  bootstrap: [AppComponent],
+
 })
 export class AppModule { }
