@@ -1,15 +1,16 @@
 import { Routes } from '@angular/router';
 
-// import { DashboardComponent } from '../../components/dashboard/dashboard.component';
-// import { UserProfileComponent } from '../../user-profile/user-profile.component';
-// import { TableListComponent } from '../../table-list/table-list.component';
-// import { TypographyComponent } from '../../typography/typography.component';
-// import { IconsComponent } from '../../icons/icons.component';
-// import { MapsComponent } from '../../maps/maps.component';
-// import { NotificationsComponent } from '../../notifications/notifications.component';
-// import { UpgradeComponent } from '../../upgrade/upgrade.component';
-
-export const AdminLayoutRoutes: Routes = [
+import { DashboardComponent } from '../dashboard/dashboard.component';
+// import { UserProfileComponent } from '../user-profile/user-profile.component';
+// import { TableListComponent } from '../table-list/table-list.component';
+// import { TypographyComponent } from '../typography/typography.component';
+// import { IconsComponent } from '../icons/icons.component';
+// import { MapsComponent } from '../maps/maps.component';
+// import { NotificationsComponent } from '../notifications/notifications.component';
+// import { UpgradeComponent } from '../upgrade/upgrade.component';
+// import { InitiateComponent } from '../initiate/initiate.component';
+ import { PersonComponent } from '../person/person.component';
+export const HomeRoutes: Routes = [
     // {
     //   path: '',
     //   children: [ {
@@ -52,6 +53,9 @@ export const AdminLayoutRoutes: Routes = [
     //         component: UpgradeComponent
     //     }]
     // }
+    
+     { path: 'person',      component: PersonComponent },
+    // { path: 'initiate',      component: InitiateComponent }, 
     // { path: 'dashboard',      component: DashboardComponent },
     // { path: 'user-profile',   component: UserProfileComponent },
     // { path: 'table-list',     component: TableListComponent },
