@@ -1,7 +1,6 @@
 import { Component, OnInit, ViewChild, AfterViewInit } from '@angular/core';
 import { Location, LocationStrategy, PathLocationStrategy, PopStateEvent } from '@angular/common';
 import 'rxjs/add/operator/filter';
-import { NavbarComponent } from '../../components/navbar/navbar.component';
 import { Router, NavigationEnd, NavigationStart } from '@angular/router';
 import { Subscription } from 'rxjs/Subscription';
 import PerfectScrollbar from 'perfect-scrollbar';
@@ -16,44 +15,49 @@ export class AdminLayoutComponent implements OnInit {
   private lastPoppedUrl: string;
   private yScrollStack: number[] = [];
 
-  constructor( public location: Location, private router: Router) {}
+  constructor( public location: Location, private router: Router) {
+      console.log('Ingreso a el adminlogin');
+  }
 
   ngOnInit() {
-      const isWindows = navigator.platform.indexOf('Win') > -1 ? true : false;
 
-      if (isWindows && !document.getElementsByTagName('body')[0].classList.contains('sidebar-mini')) {
-          // if we are on windows OS we activate the perfectScrollbar function
 
-          document.getElementsByTagName('body')[0].classList.add('perfect-scrollbar-on');
-      } else {
-          document.getElementsByTagName('body')[0].classList.remove('perfect-scrollbar-off');
-      }
-      const elemMainPanel = <HTMLElement>document.querySelector('.main-panel');
-      const elemSidebar = <HTMLElement>document.querySelector('.sidebar .sidebar-wrapper');
+       const isWindows = navigator.platform.indexOf('Win') > -1 ? true : false;
 
-      this.location.subscribe((ev:PopStateEvent) => {
-          this.lastPoppedUrl = ev.url;
-      });
-       this.router.events.subscribe((event:any) => {
-          if (event instanceof NavigationStart) {
-             if (event.url != this.lastPoppedUrl)
-                 this.yScrollStack.push(window.scrollY);
-         } else if (event instanceof NavigationEnd) {
-             if (event.url == this.lastPoppedUrl) {
-                 this.lastPoppedUrl = undefined;
-                 window.scrollTo(0, this.yScrollStack.pop());
-             } else
-                 window.scrollTo(0, 0);
-         }
-      });
-      this._router = this.router.events.filter(event => event instanceof NavigationEnd).subscribe((event: NavigationEnd) => {
-           elemMainPanel.scrollTop = 0;
-           elemSidebar.scrollTop = 0;
-      });
-      if (window.matchMedia(`(min-width: 960px)`).matches && !this.isMac()) {
-          let ps = new PerfectScrollbar(elemMainPanel);
-          ps = new PerfectScrollbar(elemSidebar);
-      }
+       if (isWindows && !document.getElementsByTagName('body')[0].classList.contains('sidebar-mini')) {
+           // if we are on windows OS we activate the perfectScrollbar function
+
+           document.getElementsByTagName('body')[0].classList.add('perfect-scrollbar-on');
+       } else {
+           document.getElementsByTagName('body')[0].classList.remove('perfect-scrollbar-off');
+       }
+       const elemMainPanel = <HTMLElement>document.querySelector('.main-panel');
+       const elemSidebar = <HTMLElement>document.querySelector('.sidebar .sidebar-wrapper');
+
+       this.location.subscribe((ev:PopStateEvent) => {
+           this.lastPoppedUrl = ev.url;
+       });
+        this.router.events.subscribe((event:any) => {
+           if (event instanceof NavigationStart) {
+              if (event.url != this.lastPoppedUrl){
+                //  this.yScrollStack.push(window.scrollY);
+             }
+          } else if (event instanceof NavigationEnd) {
+              if (event.url == this.lastPoppedUrl) {
+                  this.lastPoppedUrl = undefined;
+                 // window.scrollTo(0, this.yScrollStack.pop());
+              }// else
+                 // window.scrollTo(0, 0);
+          }
+       });
+       this._router = this.router.events.filter(event => event instanceof NavigationEnd).subscribe((event: NavigationEnd) => {
+            elemMainPanel.scrollTop = 0;
+            elemSidebar.scrollTop = 0;
+       });
+       if (window.matchMedia(`(min-width: 960px)`).matches && !this.isMac()) {
+           let ps = new PerfectScrollbar(elemMainPanel);
+           ps = new PerfectScrollbar(elemSidebar);
+       }
   }
   ngAfterViewInit() {
       this.runOnRouteChange();
@@ -69,11 +73,11 @@ export class AdminLayoutComponent implements OnInit {
       }
   }
   runOnRouteChange(): void {
-    if (window.matchMedia(`(min-width: 960px)`).matches && !this.isMac()) {
-      const elemMainPanel = <HTMLElement>document.querySelector('.main-panel');
-      const ps = new PerfectScrollbar(elemMainPanel);
-      ps.update();
-    }
+     if (window.matchMedia(`(min-width: 960px)`).matches && !this.isMac()) {
+       const elemMainPanel = <HTMLElement>document.querySelector('.main-panel');
+       const ps = new PerfectScrollbar(elemMainPanel);
+       ps.update();
+     }
   }
   isMac(): boolean {
       let bool = false;
