@@ -3,6 +3,9 @@
 // `ng build --env=prod` then `environment.prod.ts` will be used instead.
 // The list of which env maps to which file can be found in `.angular-cli.json`.
 
+
 export const environment = {
-  production: false
+  production: false,
+    scaURL: 'https://localhost/api/auth',
+    UrlApi: 'http://localhost:3000/api'
 };
